// src/enums/projectType.ts
export enum ProjectType {
  LANDING_PAGE = "landing_page",
  INSTITUTIONAL_WEBSITE = "institutional_website",
  ECOMMERCE = "ecommerce",
  BLOG = "blog",
  DASHBOARD = "dashboard",
  MOBILE_APP = "mobile_app",
  SAAS_PLATFORM = "saas_platform",
  COMPLEX_SYSTEM = "complex_system",
  API_BACKEND = "api_backend",
  OTHER = "other",
}